// src/api/generated/model/taskUpdateRequestDto.ts
// @ts-nocheck
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Dev Blackbox API
 * OpenAPI spec version: 0.1.1
 */
import type { TaskStatusEnum } from './taskStatusEnum';

export interface TaskUpdateRequestDto {
  /** 태스크 제목 */
  title: string;
  /** 태스크 내용 */
  content?: string;
  /** 태스크 태그 */
  tags?: string | null;
  /** 태스크 상태 */
  status: TaskStatusEnum;
  /** 표시 순서 */
  display_order?: number;
}